zebra: include full nexthop info in dataplane group
Include the nexthop info for every nexthop in the nexthop
group installation.

Some dataplanes may want to create the nexthop group and
every nexthop in the group with one dataplane message, due
to async issues. So we will include all of the nexthop's
info along with the weight and ID in the dataplane context
object even if we have already sent the individual singletons
before.

Signed-off-by: Stephen Worley <sworley@cumulusnetworks.com>