zebra: don't alloc group of nexthops for nh_ctx
When we are sending a single nexthop to the dataplane, don't
alloc it as if we are expecting a group. That is handled by
the nh_grp array.

We switch this to just use a single nexthop struct in
the dataplane context object rather than a nexthop_group
and handle that properly in netlink code.

Signed-off-by: Stephen Worley <sworley@cumulusnetworks.com>