lib: Fix atomic node lock warnings with enable-dev-build
The route_node lock variable type adds a const qualifier when
--enable-dev-build is configured. This is to generate warnings
in case client code tries to modify this variable. As a result
the functions route_lock_node() and route_unlock_node() need to
cast away the const qualifier to prevent the compile warnings.
When they _Atomic qualifier was added I wrongly removed the
cast. This change adds it back to prevent the warnings.

Signed-off-by: Barry Friedman (friedman) <friedman@cisco.com>