zebra: Use RTA_VIA for IPv4 hosts over IPv6 overlay forwarding
This commit enables Kernel forwarding for IPv4 hosts which are reachable
over an IPv6 overlay. The RTA_VIA attribute is used for correct
forwarding in the kernel.

LWTUNNEL Encapsulation is also updated to use the correct address family
for the underlying gateway address.

A minor bugfix, when Zebra receives a route add, the nexthop type is
used instead of the route address family in order to decode the EVPN
Nexthop Address correctly. Otherwise it tries to truncate the IPv6 VTEP
address and use it as a IPv4 VTEP...

Signed-off-by: Mike Mallin <mmallin@cisco.com>

Co-authored-by: Barry Friedman <friedman@cisco.com>