lib,zebra: replace VRF_GET_ID macro
Remove a macro and replace it with a simple function. The macro
returned from the calling function in error paths; but since the
application code wasn't involved, it wasn't able to clean up
memory allocations. Replace the macro with a simple function
that returns to the application code; the application can
then handle errors itself.

Signed-off-by: Mark Stapp <mjs@cisco.com>